The next version number will be 0.2.32.

+++ src/libawkward/array/NumpyArray.cpp
@@ -4803,7 +4803,7 @@ namespace awkward {
                                           shape,
                                           strides,
                                           0,
-                                          util::dtype_to_itemsize(dtype),
+                                          (ssize_t)util::dtype_to_itemsize(dtype),
                                           util::dtype_to_format(dtype),
                                           dtype);
     }
